backend/pyproject.toml
[project]
name = "student-system-backend"
version = "1.0.0"
description = "學生資料查詢系統後端"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "mysql-connector-python>=8.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "sqlmodel>=0.0.23"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0"
]
